cliente livro repository teste criado e repository corrigido

+++ repository/clientelivro_repository.js
@@ -10,28 +10,34 @@ const conexao = {
   password: "1"
 }
 
-async function retirarLivro(clienteId,livroId){
+async function retirarLivro(clienteId, livroId) {
   const client = new Client(conexao);
   await client.connect();
-  const livro = await livroRepository.buscarPorId(livroId);
-  const cliente = await clienteRepository.buscarPorId(clienteId)
-  if(livro &&  cliente){
-    const dataRetirada = new Date( );
-    const sql = `
+
+  const sqlCheckRelacao = `
+    SELECT * FROM cliente_livro
+    WHERE cliente_id = $1 AND livro_id = $2;
+  `;
+  const valuesCheckRelacao = [clienteId, livroId];
+  const resultCheckRelacao = await client.query(sqlCheckRelacao, valuesCheckRelacao);
+
+  if (resultCheckRelacao.rows.length === 0) {
+    const dataRetirada = new Date();
+    const sqlInsertRelacao = `
       INSERT INTO cliente_livro (cliente_id, livro_id, data_retirada)
       VALUES ($1, $2, $3)
       RETURNING *;
-      `
-      const values = [clienteId, livroId, dataRetirada];
-      const result = await client.query(sql, values);
-      const livroRetirado = result.rows[0];
-      await client.end();
-      return livroRetirado;
-      
-  }
-  else{
+    `;
+    const valuesInsertRelacao = [clienteId, livroId, dataRetirada];
+    const resultInsertRelacao = await client.query(sqlInsertRelacao, valuesInsertRelacao);
+
     await client.end();
-  }  
+
+    return resultInsertRelacao.rows[0];
+  }
+
+  await client.end();
+  return null;
 }
 
 
@@ -39,18 +45,29 @@ async function devolverLivro(clienteId, livroId) {
   const client = new Client(conexao);
   await client.connect();
 
-  // Remove a relação na tabela intermediária
-  const sqlRemoveRelacao = `
-    DELETE FROM cliente_livro
-    WHERE cliente_id = $1 AND livro_id = $2
-    RETURNING *;
+  const sqlCheckRelacao = `
+    SELECT * FROM cliente_livro
+    WHERE cliente_id = $1 AND livro_id = $2;
   `;
-  const valuesRemoveRelacao = [clienteId, livroId];
-  const resultRemoveRelacao = await client.query(sqlRemoveRelacao, valuesRemoveRelacao);
+  const valuesCheckRelacao = [clienteId, livroId];
+  const resultCheckRelacao = await client.query(sqlCheckRelacao, valuesCheckRelacao);
+
+  if (resultCheckRelacao.rows.length > 0) {
+    const sqlRemoveRelacao = `
+      DELETE FROM cliente_livro
+      WHERE cliente_id = $1 AND livro_id = $2
+      RETURNING *;
+    `;
+    const valuesRemoveRelacao = [clienteId, livroId];
+    const resultRemoveRelacao = await client.query(sqlRemoveRelacao, valuesRemoveRelacao);
+
+    await client.end();
+
+    return resultRemoveRelacao.rows[0];
+  }
 
   await client.end();
-
-  return resultRemoveRelacao.rows[0];
+  return null;
 }
